+++ pses_basis/include/pses_basis/communicationconfig.h
@@ -12,7 +12,8 @@
 #include <unordered_set>
 #include <pses_basis/syntax.h>
 
-struct SerialInterfaceParams{
+struct SerialInterfaceParams
+{
   unsigned int baudRate;
   std::string deviceTag;
   unsigned int serialTimeout;
@@ -27,24 +28,29 @@ public:
   CommunicationConfig(const CommunicationConfig& other);
   CommunicationConfig(std::string configPath);
   friend void operator>>(const YAML::Node& node, Syntax& syntax);
-  friend void operator>>(const YAML::Node& node, SerialInterfaceParams& serialParams);
   friend void operator>>(const YAML::Node& node,
-                         std::unordered_map<std::string, Channel>& channels);
+                         SerialInterfaceParams& serialParams);
+  friend void operator>>(
+      const YAML::Node& node,
+      std::unordered_map<std::string, std::shared_ptr<Channel>>& channels);
   friend void
   operator>>(const YAML::Node& node,
-             std::unordered_map<std::string, CommandOptions>& options);
+             std::unordered_map<std::string, std::shared_ptr<CommandOptions>>&
+                 options);
 
   const std::shared_ptr<Syntax> getSyntax() const;
   const std::shared_ptr<SerialInterfaceParams> getSerialInterfaceParams() const;
-  const std::unordered_map<std::string, std::shared_ptr<Command> >& getCommands() const;
-  const std::unordered_map<unsigned char, std::shared_ptr<SensorGroup> >& getSensorGroups() const;
+  const std::unordered_map<std::string, std::shared_ptr<Command>>&
+  getCommands() const;
+  const std::unordered_map<unsigned char, std::shared_ptr<SensorGroup>>&
+  getSensorGroups() const;
 
 private:
   std::string configPath;
-  std::unordered_map<std::string, std::shared_ptr<Command> > commands;
-  std::unordered_map<unsigned char, std::shared_ptr<SensorGroup> > sensorGroups;
-  std::unordered_map<std::string, Channel> channels;
-  std::unordered_map<std::string, CommandOptions> options;
+  std::unordered_map<std::string, std::shared_ptr<Command>> commands;
+  std::unordered_map<unsigned char, std::shared_ptr<SensorGroup>> sensorGroups;
+  std::unordered_map<std::string, std::shared_ptr<Channel>> channels;
+  std::unordered_map<std::string, std::shared_ptr<CommandOptions>> options;
   Syntax syntax;
   SerialInterfaceParams serialParams;
